moves files to packages, updates tests
