+++ backend/src/main/java/org/example/backend/player/PlayerRepo.java
@@ -1,4 +1,4 @@
-package org.example.backend;
+package org.example.backend.player;
 
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
