+++ backend/src/main/java/org/example/backend/puzzle/Puzzle.java
@@ -1,4 +1,4 @@
-package org.example.backend;
+package org.example.backend.puzzle;
 
 public record Puzzle(String puzzleId, String title, String body, String solution) {
 }
