+++ backend/src/main/java/org/example/backend/puzzle/PuzzleController.java
@@ -1,4 +1,4 @@
-package org.example.backend;
+package org.example.backend.puzzle;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.GetMapping;
