+++ backend/src/main/java/org/example/backend/puzzle/PuzzleRepo.java
@@ -1,4 +1,4 @@
-package org.example.backend;
+package org.example.backend.puzzle;
 
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
