+++ backend/src/main/java/org/example/backend/puzzle/PuzzleService.java
@@ -1,8 +1,7 @@
-package org.example.backend;
+package org.example.backend.puzzle;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
-
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Random;
