+++ backend/src/test/java/org/example/backend/player/PlayerControllerTest.java
@@ -1,4 +1,4 @@
-package org.example.backend;
+package org.example.backend.player;
 
 import lombok.RequiredArgsConstructor;
 import org.junit.jupiter.api.Test;
