+++ backend/src/test/java/org/example/backend/player/PlayerServiceTest.java
@@ -1,11 +1,9 @@
-package org.example.backend;
+package org.example.backend.player;
 
 import org.junit.jupiter.api.Test;
-
 import java.util.List;
 import java.util.NoSuchElementException;
 import java.util.Optional;
-
 import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.Mockito.*;
